tests: move python supporting functions to a {scripts} directory
